requirements.txt
backcall==0.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.0
Flask==1.0.2
idna==2.8
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jupyter-client==5.2.4
jupyter-core==4.4.0
MarkupSafe==1.1.0
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
ptyprocess==0.6.0
Pygments==2.3.0
python-dateutil==2.7.5
python-dotenv==0.10.1
pyzmq==17.1.2
requests==2.21.0
responses==0.10.4
six==1.12.0
tornado==5.1.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
